Moving map card components to app dir
