+++ src/app/(private)/components/MapCard.tsx
